+++ app/helpers/linebot_helper.rb
@@ -1,0 +1,2 @@
+module LinebotHelper
+end
