+++ app/views/items/index.html.erb
@@ -24,6 +24,9 @@
               <div class="absolute top-2 right-2 flex space-x-2">
                 <div class="flex justify-end space-x-6">
                   <%= link_to edit_item_path(item), class: "text-gray-500 hover:text-gray-700" do %>
+                    <i class="fas fa-clock text-3xl"></i>
+                  <% end %>
+                  <%= link_to edit_item_path(item), class: "text-gray-500 hover:text-gray-700" do %>
                     <i class="fas fa-pen text-3xl"></i>
                   <% end %>
                   <%= link_to item_path(item), method: :delete, data: { turbo_method: :delete, turbo_confirm: "本当に削除しますか？" }, class: "text-gray-500 hover:text-gray-700" do %>
