+++ app/views/notification/edit.html.erb
@@ -1,0 +1,1 @@
+作成中
